Fix amax device, fp32 bias dtype, and the cutlass-dsl version probe
- Move global_amax to x's device in the CuTe DSL quantize entry: the
  kernel reads it through a raw device pointer, so the CPU scalar that
  works on the pure-PyTorch body handed the kernel a host pointer.
- Cast an fp32 bias to the bf16 output dtype in four_over_six_mm so the
  layer output dtype is not silently promoted.
- Tolerate PackageNotFoundError from the nvidia-cutlass-dsl version
  lookup (cutlass can ship under another distribution name) instead of
  crashing every moe_training import.

+++ torchao/prototype/moe_training/nvfp4_training/four_over_six.py
@@ -529,7 +529,7 @@ class four_over_six_mm(torch.autograd.Function):
             )
         output = output.reshape(*input_hp.shape[:-1], N)
         if bias is not None:
-            output = output + bias
+            output = output + bias.to(output.dtype)
 
         if backward_override == "high_precision":
             ctx.save_for_backward(input_2d, weight_hp)
